requirements.txt
crewai==0.11.0
langchain>=0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
streamlit==1.29.0
python-dotenv==1.0.0
duckduckgo-search==4.1.1